refactor: restructure component packages to 'ui.movio_component'
