+++ ui/src/main/java/com/cairosquad/ui/search/content/SearchContent.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.cairosquad.design_system.R
 import com.cairosquad.design_system.component.InputField
-import com.cairosquad.design_system.component.RecentSearchItem
-import com.cairosquad.design_system.component.SectionHeader
+import com.cairosquad.ui.movio_component.RecentSearchItem
+import com.cairosquad.ui.movio_component.SectionHeader
 import com.cairosquad.design_system.theme.Theme
 import com.cairosquad.viewmodel.search.SearchInteractionListener
 import com.cairosquad.viewmodel.search.SearchScreenState
