+++ ui/src/main/java/com/cairosquad/ui/search/content/SearchScreenContent.kt
@@ -4,7 +4,7 @@ import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.cairosquad.design_system.component.RefreshBox
+import com.cairosquad.ui.movio_component.RefreshBox
 import com.cairosquad.viewmodel.search.SearchInteractionListener
 import com.cairosquad.viewmodel.search.SearchScreenState
